Chore: add thin top.handlers shim to controller; Guardrail: general *.legacy.js exclusion
